Update decaf library
Generate C code from upstream commit
cede6185df1dad9d6e465dddc7f02698d8553726

The following code in the upstream repo was run:
    cmake .
    make decaf_tables decaf-static all

+++ cbits/decaf/p448/arch_32/f_impl.c
@@ -88,11 +88,11 @@ void crypton_gf_mulw_unsigned (crypton_gf_s *__restrict__ cs, const gf as, uint3
 
     accum0 += accum8 + c[8];
     c[8] = accum0 & mask;
-    c[9] += accum0 >> 28;
+    c[9] += (uint32_t)(accum0 >> 28);
 
     accum8 += c[0];
     c[0] = accum8 & mask;
-    c[1] += accum8 >> 28;
+    c[1] += (uint32_t)(accum8 >> 28);
 }
 
 void crypton_gf_sqr (crypton_gf_s *__restrict__ cs, const gf as) {
